+++ fhir_core/xml_utils.py
@@ -765,7 +765,8 @@ class Node:
                 # we filter non-summary field
                 continue
             field_ = value.__class__.model_fields[alias_maps[prop_name]]
-            val = value.__dict__.get(field_.alias)
+            field_key_ = alias_maps[field_.alias]
+            val = value.__dict__.get(field_key_)
             if fhir_type_name == "Extension" and field_.alias in ("url", "id") and val:
                 child.add_attribute(field_.alias, val)
                 continue
@@ -783,7 +784,7 @@ class Node:
 
             value_ext, value_ext_field = None, None
             if is_primitive_type(field_):
-                ext_key = f"{alias_maps[field_.alias]}__ext"
+                ext_key = f"{field_key_}__ext"
                 value_ext = value.__dict__.get(ext_key, None)
                 if value_ext:
                     value_ext_field = value.__class__.model_fields[ext_key]
@@ -819,13 +820,14 @@ class Node:
             ):
                 # we filter a non-summary element
                 continue
-            field = model.__class__.model_fields[alias_maps[prop_name]]
+            field_key = alias_maps[prop_name]
+            field = model.__class__.model_fields[field_key]
             if typing.TYPE_CHECKING:
                 assert field.alias
-            value = model.__dict__.get(field.alias, None)
+            value = model.__dict__.get(field_key, None)
             value_ext, value_ext_field = None, None
             if is_primitive_type(field):
-                ext_key = f"{field.alias}__ext"
+                ext_key = f"{field_key}__ext"
                 value_ext = model.__dict__.get(ext_key, None)
                 if value_ext:
                     value_ext_field = model.__class__.model_fields[ext_key]